refactor: use new interfaces
* remove temporary 'WithInterfaces' classes
