+++ src/main/java/org/janelia/saalfeldlab/n5/N5FSWriter.java
@@ -36,7 +36,7 @@ import com.google.gson.GsonBuilder;
  * @author Stephan Saalfeld
  */
 //public class N5FSWriter extends N5KeyValueWriter implements N5Writer {
-public class N5FSWriter extends N5KeyValueWriterWithInterfaces {
+public class N5FSWriter extends N5KeyValueWriter {
 
 	/**
 	 * Opens an {@link N5FSWriter} at a given base path with a custom
